Adjust river layout and called tile rotation
